Fix bugs and update AndroidCommon

+++ androidCommon/src/main/java/com/asksven/android/common/kernelutils/NativeKernelWakelock.java
@@ -16,31 +16,19 @@
 
 package com.asksven.android.common.kernelutils;
 
-import java.io.Serializable;
-import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.List;
-import java.lang.Math;
+import android.util.Log;
 
-import com.asksven.android.common.dto.AlarmDto;
 import com.asksven.android.common.dto.NativeKernelWakelockDto;
-import com.asksven.android.common.nameutils.UidInfo;
 import com.asksven.android.common.nameutils.UidNameResolver;
-import com.asksven.android.common.privateapiproxies.AlarmItem;
 import com.asksven.android.common.privateapiproxies.StatElement;
 import com.asksven.android.common.utils.StringUtils;
 import com.google.gson.annotations.SerializedName;
 
+import java.io.Serializable;
+import java.util.Comparator;
+import java.util.List;
 
-
-
-
-
-
-
-import android.os.Build;
 //import android.content.Context;
-import android.util.Log;
 
 /**
  * @author sven
@@ -432,8 +420,8 @@ public class NativeKernelWakelock extends StatElement implements Comparable<Nati
 	{
 		public int compare(NativeKernelWakelock a, NativeKernelWakelock b)
 		{
-			return ((int)(b.getCount() - a.getCount()));
-		}
+            return b.getCount() - a.getCount();
+        }
 	}
 	
 	public static class TimeComparator implements Comparator<NativeKernelWakelock>
