+++ androidCommon/src/main/java/com/asksven/android/common/privateapiproxies/KernelWakelock.java
@@ -16,12 +16,13 @@
 
 package com.asksven.android.common.privateapiproxies;
 
+import android.util.Log;
+
+import com.google.gson.annotations.SerializedName;
+
 import java.io.Serializable;
 import java.util.Comparator;
 import java.util.List;
-import com.google.gson.annotations.SerializedName;
-
-import android.util.Log;
 
 
 /**
@@ -181,8 +182,8 @@ public class KernelWakelock extends StatElement implements Comparable<KernelWake
 	{
 		public int compare(KernelWakelock a, KernelWakelock b)
 		{
-			return ((int)(b.getCount() - a.getCount()));
-		}
+            return b.getCount() - a.getCount();
+        }
 	}
 	
 	public static class TimeComparator implements Comparator<KernelWakelock>
